% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{litterGeeDf}
\alias{litterGeeDf}
\title{Small litter data set with overdispersion}
\format{A data frame with 80 rows and 3 variables:
\describe{
  \item{treat}{Treatment (A, B, C, and D)}
  \item{litter}{Litter ID}
  \item{resp}{Count response of a treatment}
}}
\source{
Simulated data set
}
\usage{
litterGeeDf
}
\description{
A dataset containing a treatment, litter, and count column
}
\keyword{datasets}

